voice/requirements.txt
python-dotenv
aiohttp
uvloop
logger